requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
APScheduler==3.10.4
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.1
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.8.4
multidict==6.6.4
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.2
packaging==25.0
pandas==2.2.2
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
StrEnum==0.4.15
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
